13.1.2-2 baseline
Former-commit-id: 51208e55154a3a494cd4ae04b4de90919cdf7c71 [formerly c0c1d56f3aa2efe0f6825d70ff9e56407d690836] [formerly 4ba416aeb32ba04688841ada586ed9d1be6660f4 [formerly 221d0aaf0e19a054387134e72a76a278159bf4d0]]
Former-commit-id: 4ba416aeb32ba04688841ada586ed9d1be6660f4
Former-commit-id: 941337c5538b8031bb492c28cceecbf7759798be

+++ edexOsgi/com.raytheon.edex.plugin.shef/unit-test/test/edex/transform/shef/TestM2SOptions.java
@@ -19,12 +19,15 @@
  **/
 package test.edex.transform.shef;
 
-import org.junit.Test;
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.StringReader;
+
+import org.junit.Test;
 
 import com.raytheon.edex.transform.shef.obs.ObsToSHEFOptions;
 
@@ -86,7 +89,7 @@ public class TestM2SOptions {
             }
         }
 
-        assertEquals(40, options.getPCReset("KOMA"));
+        assertEquals(40, options.getPCReset("KOMA").intValue());
         assertFalse(options.checkName("KSAT"));
         assertTrue(options.isOptZeroAuto1HourPrecip());
         assertFalse(options.isOptTypeSrcV());
